// backend/src/domain/types.ts
export type SubjectCode =
    | "PORT"
    | "ENGL"
    | "ART"
    | "EDF"
    | "MAT"
    | "HIS"
    | "GEO"
    | "SOC"
    | "FIL"
    | "FIS"
    | "QUI"
    | "BIO"

export type SubjectName =
    | "Lingua Portuguesa"
    | "Lingua Inglesa"
    | "Artes"
    | "Educacao Fisica"
    | "Matematica"
    | "Historia"
    | "Geografia"
    | "Sociologia"
    | "Filosofia"
    | "Fisica"
    | "Quimica"
    | "Biologia"